Function arrayfire::bilateral

pub fn bilateral<T>(
    input: &Array<T>, 
    spatial_sigma: f32, 
    chromatic_sigma: f32, 
    iscolor: bool
) -> Array<T::AbsOutType> where
    T: HasAfEnum + ImageFilterType,
    T::AbsOutType: HasAfEnum, 

Bilateral Filter.

A bilateral filter is a edge-preserving filter that reduces noise in an image. The intensity of each pixel is replaced by a weighted average of the intensities of nearby pixels. The weights follow a Gaussian distribution and depend on the distance as well as the color distance.

The bilateral filter requires the size of the filter (in pixels) and the upper bound on color values, N, where pixel values range from 0–N inclusively.

Parameters

input array is the input image

spatial_sigma is the spatial variance parameter that decides the filter window

chromatic_sigma is the chromatic variance parameter

iscolor indicates if the input is color image or grayscale

Return Values

Filtered Image - Array