Function arrayfire::cholesky

pub fn cholesky(input: &Array, is_upper: bool) -> Result<(Array, i32), AfError>

Perform Cholesky decomposition

Parameters

input is the input matrix

is_upper is a boolean to indicate if the output has to be upper or lower triangular matrix

Return Values

A tuple of an Array and signed 32-bit integer.

The Array contains the triangular matrix (multiply it with conjugate transpose to reproduce the input).

If the integer is 0, it means the cholesky decomposition passed. Otherwise, it will contain the rank at which the decomposition failed.