Function arrayfire::constant_t

pub fn constant_t(value: Scalar, dims: Dim4, dtype: DType) -> Array

Create a constant array of given type(DType)

You can use this function to create arrays of type dictated by the enum DType using the scalar value that has the shape similar to dims.

Parameters

value is the Scalar to be filled into the array

dims is the output Array dimensions

dtype indicates the type of Array to be created and is the type of the scalar to be passed via the paramter value.

Return Values

Array of dims shape and filed with given constant value.