Function arrayfire::convolve2_gradient_nn

pub fn convolve2_gradient_nn<T>(
    incoming_grad: &Array<T>, 
    original_signal: &Array<T>, 
    original_filter: &Array<T>, 
    convolved_output: &Array<T>, 
    strides: Dim4, 
    padding: Dim4, 
    dilation: Dim4, 
    grad_type: ConvGradientType
) -> Array<T> where
    T: HasAfEnum + RealFloating, 

Backward pass gradient of 2D convolution

Parameters

incoming_gradient gradients to be distributed in backwards pass

original_signal input signal to forward pass of convolution assumed structure of input is ( d0 x d1 x d2 x N )

original_filter input filter to forward pass of convolution assumed structure of input is ( d0 x d1 x d2 x N )

convolved_output output from forward pass of convolution

strides are distance between consecutive elements along each dimension for original convolution

padding specifies padding width along each dimension for original convolution

dilation specifies filter dilation along each dimension for original convolution

grad_type specifies which gradient to return

Return Values

Gradient Array w.r.t input generated from convolve2_nn