Function arrayfire::convolve2_sep

pub fn convolve2_sep<T, F>(
    cfilt: &Array<F>, 
    rfilt: &Array<F>, 
    signal: &Array<T>, 
    mode: ConvMode
) -> Array<T> where
    T: HasAfEnum,
    F: HasAfEnum, 

Separable convolution for 2d signals

Parameters

cfilt is the filter to be applied along coloumns

rfilt is the filter to be applied along rows

signal is the input signal

mode indicates if the convolution should be expanded or not(where output size equals input)

Return Values

The convolved Array