Function arrayfire::dot

pub fn dot(lhs: &Array, rhs: &Array, optlhs: MatProp, optrhs: MatProp) -> Result<Array, AfError>

Calculate the dot product of vectors.

Scalar dot product between two vectors. Also referred to as the inner product. This function returns the scalar product of two equal sized vectors or between a matrix and a vector. The second operand needs to be a vector in either case.

Parameters

lhs - Left hand side of dot operation

rhs - Right hand side of dot operation

optlhs - Options for lhs. Currently only NONE value from MatProp is supported.

optrhs - Options for rhs. Currently only NONE value from MatProp is supported.

Return Values

The result of dot product.