Function arrayfire::erode

pub fn erode<T>(input: &Array<T>, mask: &Array<T>) -> Array<T> where
    T: HasAfEnum + ImageFilterType, 

Erode an Image

The erosion function is a morphological transformation on an image that requires two inputs. The first is the image to be morphed, and the second is the mask indicating neighborhood that must be white in order to preserve each pixel.

In erode, for each pixel, the mask is centered at the pixel. If each pixel of the mask matches the corresponding pixel on the image, then no change is made. If there is at least one mismatch, then pixels are changed to the background color (black).

For further reference, see here.

Parameters

input is the input image

mask is the morphological operation mask

Return Values

Eroded Image(Array)