Function arrayfire::fft2

pub fn fft2(input: &Array, norm_factor: f64, odim0: i64, odim1: i64) -> Result<Array, AfError>

Fast fourier transform for 2d signals

Parameters

input is the input Array

norm_factor is the normalization factor with which the input is scaled before the transformation is applied

odim0 is the length of output signal first dimension - used for either truncating or padding the input

odim1 is the length of output signal second dimension - used for either truncating or padding the input

Return Values

Transformed Array