Function arrayfire::fft_convolve2

pub fn fft_convolve2(signal: &Array, filter: &Array, mode: ConvMode) -> Array

Convolution using Fast-fourier transform

The numeric suffix to the function name indicates the dimension in which the convolution operation is going to take place.

1 - Indicates 1d convolution

2 - Indicates 2d convolution

3 - Indicates 3d convolution

Parameters

signal is the input signal

filter is the signal that shall be used for convolution operation

mode indicates if the convolution should be expanded or not(where output size equals input)

Return Values

The convolved Array