Function arrayfire::ge

pub fn ge<T, U>(arg1: &T, arg2: &U, batch: bool) -> Array<bool> where
    T: Convertable,
    U: Convertable,
    <T as Convertable>::OutType: HasAfEnum + ImplicitPromote<<U as Convertable>::OutType>,
    <U as Convertable>::OutType: HasAfEnum + ImplicitPromote<<T as Convertable>::OutType>, 

Perform greater than equals comparison operation

This is a comparison operation.

Parameters

arg1is an argument that implements an internal trait Convertable.

arg2is an argument that implements an internal trait Convertable.

batch is an boolean that indicates if the current operation is an batch operation.

Both parameters arg1 and arg2 can be either an Array or a value of rust integral type.

Return Values

An Array with results of the comparison operation a.k.a an Array of boolean values.

Important Notes

If shape/dimensions of arg1 and arg2 are same, the value of batch parameter has no effect.

If shape/dimensions of arg1 and arg2 are different, the value of batch has to be set to true. In this case, the shapes of arg1 and arg2 have to satisfy the following criteria:

Same number of elements in arg1 and arg2 along a given dimension/axis

Only one element in arg1 or arg2 along a given dimension/axis

The trait Convertable essentially translates to a scalar native type on rust or Array.