Function arrayfire::hamming_matcher

pub fn hamming_matcher(query: &Array, train: &Array, dist_dims: i64, n_dist: u32) -> (Array, Array)

Hamming feature matcher

Calculates Hamming distances between two 2-dimensional arrays containing features, one of the arrays containing the training data and the other the query data. One of the dimensions of the both arrays must be equal among them, identifying the length of each feature. The other dimension indicates the total number of features in each of the training and query arrays. Two 1-dimensional arrays are created as results, one containg the smallest N distances of the query array and another containing the indices of these distances in the training array. The resulting 1-dimensional arrays have length equal to the number of features contained in the query array.

Parameters

query - Array containing the data to be queried

train - Array containing the data to be used as training data

dist_dims - indicates the dimension to analyze for distance (the dimension indicated here must be of equal length for both query and train arrays)

n_dist - is the number of smallest distances to return (currently, only 1 is supported)

Return Values

This function returns a tuple of Array's.

First Array is an array of MxN size, where M is equal to the number of query features and N is equal to n_dist. The value at position IxJ indicates the index of the Jth smallest distance to the Ith query value in the train data array. the index of the Ith smallest distance of the Mth query.

Second Array is an array of MxN size, where M is equal to the number of query features and N is equal to n_dist. The value at position IxJ indicates the Hamming distance of the Jth smallest distance to the Ith query value in the train data array.