Function arrayfire::histogram

pub fn histogram<T>(
    input: &Array<T>, 
    nbins: u32, 
    minval: f64, 
    maxval: f64
) -> Array<u32> where
    T: HasAfEnum + RealNumber, 

Compute Histogram of an Array

A histogram is a representation of the distribution of given data. This representation is essentially a graph consisting of the data range or domain on one axis and frequency of occurence on the other axis. All the data in the domain is counted in the appropriate bin. The total number of elements belonging to each bin is known as the bin’s frequency.

The regular histogram function creates bins of equal size between the minimum and maximum of the input data (min and max are calculated internally). The histogram min-max function takes input parameters minimum and maximum, and divides the bins into equal sizes within the range specified by min and max parameters. All values less than min in the data range are placed in the first (min) bin and all values greater than max will be placed in the last (max) bin.

Parameters

input is the Array whose histogram has to be computed

nbins is the number bins the input data has to be categorized into.

minval is the minimum value of bin ordering

maxval is the maximum value of bin ordering

Return Values

Histogram of input Array