Function arrayfire::iota

pub fn iota<T: HasAfEnum>(dims: Dim4, tdims: Dim4) -> Array<T>

Create a range of values

Create an sequence [0, dims.elements() - 1] and modify to specified dimensions dims and then tile it according to tile_dims.

Parameters

dims is the dimensions of the sequence to be generated

tdims is the number of repitions of the unit dimensions

Return Values

Array