Function arrayfire::join_many

pub fn join_many<T>(dim: i32, inputs: Vec<&Array<T>>) -> Array<T> where
    T: HasAfEnum, 

Join multiple arrays

Parameters

dim is the dimension along which the concatenation has to be done

inputs is the vector of Arrays that has to be concatenated

Return Values

Concatenated Array