Function arrayfire::load_image_native

pub fn load_image_native<T>(filename: String) -> Array<T> where
    T: HasAfEnum + ImageNativeType, 

Load Image into Array in it's native type

This load image function allows you to load images as U8, U16 or F32 depending on the type of input image as shown by the table below.

Bits per Color (Gray/RGB/RGBA Bits Per Pixel)	Array Type	Range

8 ( 8/24/32 BPP)	u8	0 - 255

16 (16/48/64 BPP)	u16	0 - 65535

32 (32/96/128 BPP)	f32	0 - 1

Parameters

filename is name of file to be loaded

Return Arrays

An Array with pixel values loaded from the image