Function arrayfire::lu

pub fn lu(input: &Array) -> (Array, Array, Array)

Perform LU decomposition

Parameters

input is the input matrix

Return Values

A triplet of Arrays.

The first Array will contain the lower triangular matrix of the LU decomposition.

The second Array will contain the lower triangular matrix of the LU decomposition.

The third Array will contain the permutation indices to map the input to the decomposition.