Function arrayfire::lu_inplace

pub fn lu_inplace(input: &mut Array, is_lapack_piv: bool) -> Result<Array, AfError>

Perform inplace LU decomposition

Parameters

input contains the input matrix on entry and packed LU decomposition on exit

is_lapack_pic specified if the pivot is returned in original LAPACK compliant format

Return Values

An Array with permutation indices to map the input to the decomposition. Since, the input matrix is modified in place, only pivot values are returned.