Function arrayfire::match_template

pub fn match_template<T>(
    search_img: &Array<T>, 
    template_img: &Array<T>, 
    mtype: MatchType
) -> Array<T::AbsOutType> where
    T: HasAfEnum + ImageFilterType,
    T::AbsOutType: HasAfEnum, 

Image matching

Template matching is an image processing technique to find small patches of an image which match a given template image. A more in depth discussion on the topic can be found here.

Parameters

search_img is an array with image data

template_img is the template we are looking for in the image

mtype is metric that should be used to calculate the disparity between window in the image and the template image. It can be one of the values defined by the enum MatchType.

Return Values

This function returns an Array with disparity values for the window starting at corresponding pixel position.