Function arrayfire::matmul

pub fn matmul(lhs: &Array,
              rhs: &Array,
              optlhs: MatProp,
              optrhs: MatProp)
              -> Array

Matrix multiple of two Arrays

Parameters

lhs is the Array on left hand side

rhs is the Array on right hand side

optlhs - Transpose left hand side before the function is performed, uses one of the values of MatProp

optrhs - Transpose right hand side before the function is performed, uses one of the values of MatProp

Return Values

The result Array of matrix multiplication