Function arrayfire::mean_shift

pub fn mean_shift(input: &Array, spatial_sigma: f32, chromatic_sigma: f32, iter: u32, iscolor: bool) -> Result<Array, AfError>

Meanshift Filter.

A meanshift filter is an edge-preserving smoothing filter commonly used in object tracking and image segmentation.

This filter replaces each pixel in the image with the mean of the values within a given given color and spatial radius. The meanshift filter is an iterative algorithm that continues until a maxium number of iterations is met or until the value of the means no longer changes.

Parameters

input array is the input image

spatial_sigma is the spatial variance parameter that decides the filter window

chromatic_sigma is the chromatic variance parameter

iter is the number of iterations filter operation is performed

iscolor indicates if the input is color image or grayscale

Return Values

Filtered Image - Array