Function arrayfire::min_by_key

pub fn min_by_key<KeyType, ValueType>(
    keys: &Array<KeyType>, 
    vals: &Array<ValueType>, 
    dim: i32
) -> (Array<KeyType>, Array<ValueType::AggregateOutType>) where
    KeyType: ReduceByKeyInput,
    ValueType: HasAfEnum,
    ValueType::AggregateOutType: HasAfEnum, 

Find minimum among values of similar keys along a given dimension

Parameters

keys - key Array

vals - value Array

dim - Dimension along which the input Array is reduced

Return Values

Tuple of Arrays, with output keys and values after reduction