Function arrayfire::minof

pub fn minof<A, B>(
    lhs: &Array<A>, 
    rhs: &Array<B>, 
    batch: bool
) -> Array<A::Output> where
    A: ImplicitPromote<B>,
    B: ImplicitPromote<A>, 

Elementwise minimum operation of two Arrays

This is an element wise binary operation.

Important Notes

If shape/dimensions of lhs and rhs are same, the value of batch parameter has no effect.

If shape/dimensions of lhs and rhs are different, the value of batch has to be set to true. In this case, the shapes of lhs and rhs have to satisfy the following criteria:

Same number of elements in lhs and rhs along a given dimension/axis

Only one element in lhs or rhs along a given dimension/axis