Function arrayfire::print_mem_info

pub fn print_mem_info(msg: String, device: i32)

Print buffer details from the ArrayFire device manager

This information is printed in the form of a table.

Parameters

msg is a message to print before the table

device is the id of the device for which buffer details are to be printed

Return Values

None