Function arrayfire::product_nan

pub fn product_nan<T>(
    input: &Array<T>, 
    dim: i32, 
    nanval: f64
) -> Array<T::ProductOutType> where
    T: HasAfEnum,
    T::ProductOutType: HasAfEnum, 

Product of elements along specific dimension using user specified value instead of NAN values

Compute product of the values of the input Array along dim dimension after replacing any NAN values in the Array with nanval value.

Parameters

input is the input Array

dim is reduction dimension

nanval is value with which all the NAN values of Array are replaced with

Return Values

Array that is reduced along given dimension via multiplication operation