Function arrayfire::product_nan_all

pub fn product_nan_all<T>(
    input: &Array<T>, 
    val: f64
) -> (<<T as HasAfEnum>::ProductOutType as HasAfEnum>::BaseType, <<T as HasAfEnum>::ProductOutType as HasAfEnum>::BaseType) where
    T: HasAfEnum,
    <T as HasAfEnum>::ProductOutType: HasAfEnum,
    <<T as HasAfEnum>::ProductOutType as HasAfEnum>::BaseType: HasAfEnum + Fromf64, 

Product of all values using user provided value for NAN

Compute the product of all the values of the input Array after replacing any NAN values with val

Parameters

input is the input Array

val is the val that replaces all NAN values of the Array before reduction operation is performed.

Return Values

A tuple of product result.

Note: For non-complex data type Arrays, second value of tuple is zero.