Function arrayfire::resize

pub fn resize<T: HasAfEnum>(
    input: &Array<T>, 
    odim0: i64, 
    odim1: i64, 
    method: InterpType
) -> Array<T>

Resize an Image

Resizing an input image can be done using either NEAREST or BILINEAR interpolations. Nearest interpolation will pick the nearest value to the location, whereas bilinear interpolation will do a weighted interpolation for calculate the new size.

This function does not differentiate between images and data. As long as the array is defined and the output dimensions are not 0, it will resize any type or size of array.

Parameters

input is the image to be resized

odim0 is the output height

odim1 is the output width

method indicates which interpolation method to use for resizing. It uses enum InterpType to identify the interpolation method.

Return Values

Resized Array