Function arrayfire::rgb2ycbcr

pub fn rgb2ycbcr(input: &Array, standard: YCCStd) -> Array

RGB to YCbCr colorspace converter.

RGB (Red, Green, Blue) is the most common format used in computer imaging. RGB stores individual values for red, green and blue, and hence the 3 values per pixel. A combination of these three values produces the gamut of unique colors.

YCbCr is a family of color spaces used as a part of the color image pipeline in video and digital photography systems where Y is luma component and Cb & Cr are the blue-difference and red-difference chroma components.

Input array to this function should be of real data in the range [0,1].

Parameters

input is the input image in RGB color space

standard is the target color space - YCbCr standard

Return Values

Image(Array) in YCbCr color space