Function arrayfire::scan

pub fn scan<T>(
    input: &Array<T>, 
    dim: i32, 
    op: BinaryOp, 
    inclusive: bool
) -> Array<T::AggregateOutType> where
    T: HasAfEnum,
    T::AggregateOutType: HasAfEnum, 

Generalized scan

Parameters

input is the data on which scan is to be performed

dim is the dimension along which scan operation is to be performed

op takes value of BinaryOp enum indicating the type of scan operation

inclusive says if inclusive/exclusive scan is to be performed

Return Values

Output Array of scanned input