Function arrayfire::shiftr

pub fn shiftr<T, U>(
    arg1: &T, 
    arg2: &U, 
    batch: bool
) -> Array<<<T as Convertable>::OutType as ImplicitPromote<<U as Convertable>::OutType>>::Output> where
    T: Convertable,
    U: Convertable,
    <T as Convertable>::OutType: HasAfEnum + ImplicitPromote<<U as Convertable>::OutType>,
    <U as Convertable>::OutType: HasAfEnum + ImplicitPromote<<T as Convertable>::OutType>,
    <<T as Convertable>::OutType as ImplicitPromote<<U as Convertable>::OutType>>::Output: HasAfEnum, 

Compute right shift

This is a binary elementwise operation.

Parameters

arg1is an argument that implements an internal trait Convertable.

arg2is an argument that implements an internal trait Convertable.

batch is an boolean that indicates if the current operation is an batch operation.

Both parameters arg1 and arg2 can be either an Array or a value of rust integral type.

Return Values

An Array with results of the binary operation.

In the case of comparison operations such as the following, the type of output Array is DType::B8. To retrieve the results of such boolean output to host, an array of 8-bit wide types(eg. u8, i8) should be used since ArrayFire's internal implementation uses char for boolean.

gt

lt

ge

le

eq

Note

The trait Convertable essentially translates to a scalar native type on rust or Array.