Function arrayfire::wrap

pub fn wrap<T: HasAfEnum>(
    input: &Array<T>, 
    ox: i64, 
    oy: i64, 
    wx: i64, 
    wy: i64, 
    sx: i64, 
    sy: i64, 
    px: i64, 
    py: i64, 
    is_column: bool
) -> Array<T>

Converts unwrapped image to an image

Wrap takes an unwrapped image (see unwrap()) and converts it back to an image.

The inputs to this function should be the same as the inputs used to generate the unwrapped image.

Parameters

input is the output of unwrap function call

ox is the 0th-dimension of output image

oy is the 1st-dimension of output image

wx is the block window size along 0th-dimension between

wy is the block window size along 1st-dimension between

sx is the stride along 0th-dimension

sy is the stride along 1st-dimension

px is the padding used along 0th-dimension between [0, wx).

py is the padding used along 1st-dimension between [0, wy).

is_column specifies the layout for the unwrapped patch. If is_column is false, the rows are treated as the patches

Return Values

Image(Array) created from unwrapped Image(Array)