Function arrayfire::ycbcr2rgb

pub fn ycbcr2rgb(input: &Array, standard: YCCStd) -> Result<Array, AfError>

YCbCr to RGB colorspace converter.

YCbCr is a family of color spaces used as a part of the color image pipeline in video and digital photography systems where Y is luma component and Cb & Cr are the blue-difference and red-difference chroma components.

RGB (Red, Green, Blue) is the most common format used in computer imaging. RGB stores individual values for red, green and blue, and hence the 3 values per pixel. A combination of these three values produces the gamut of unique colors.

Input array to this function should be of real data with the following range in their respective channels.

Y −> [16,219]

Cb −> [16,240]

Cr −> [16,240]

Parameters

input is the input image in YCbCr color space

standard is the YCbCr standard in which input image color space is present.

Return Values

Image(Array) in RGB color space