Macro arrayfire::view

macro_rules! view {
    (@ af_max_dims) => { ... };
    ($array_ident : ident) => { ... };
    ($array_ident : ident
 [$($start : literal : $end : literal : $step : literal), +]) => { ... };
    (@ set_indexer $idim : expr, $idxr : ident, $lterm : expr) => { ... };
    (@ set_indexer $idim : expr, $idxr : ident, $hterm : expr, $($tterm : expr),
 *) => { ... };
    ($array_ident : ident [$($_e : expr), +]) => { ... };
}

Indexing into an existing Array

This macro call with return an Array that has a view of another Array. The Array returned due to the indexing operation will follow copy-on-write semantics. The Array identifier taken by this macro is passed to the relevant internal functions as a borrowed reference. Thus, this identifier will be still available for futher use after the macro call.

The following types of inputs are matched by this macro.

A simple Array identifier.

An Array with slicing info for indexing.

An Array with slicing info and other arrays used for indexing.

Examples on how to use this macro are provided in the tutorials book