Struct arrayfire::RandomEngine

pub struct RandomEngine { /* fields omitted */ }

Random number generator engine

This is a wrapper for ArrayFire's native random number generator engine.

Sharing Across Threads

While sharing this object with other threads, there is no need to wrap this in an Arc object unless only one such object is required to exist. The reason being that ArrayFire's internal details that are pointed to by the RandoMEngine handle are appropriately reference counted in thread safe manner. However, if you need to modify RandomEngine object, then please do wrap the object using a Mutex or Read-Write lock.

Implementations

impl RandomEngine

pub fn new(rengine: RandomEngineType, seed: Option<u64>) -> Self

Create a new random engine object

Parameters

rengine can be value of RandomEngineType enum.

seed is the initial seed value

Return Values

A object of type RandomEngine

pub fn get_type(&self) -> RandomEngineType

Get random engine type

pub fn set_type(&mut self, engine_type: RandomEngineType)

Get random engine type

pub fn set_seed(&mut self, seed: u64)

Set seed for random engine

pub fn get_seed(&self) -> u64

Get seed of the random engine

pub unsafe fn get(&self) -> af_random_engine

Returns the native FFI handle for Rust object RandomEngine