Trait arrayfire::ConstGenerator

pub trait ConstGenerator: HasAfEnum {
    type OutType: HasAfEnum;
    fn generate(&self, dims: Dim4) -> Array<Self::OutType>;
}

Type Trait to generate a constant Array of given size

Internally, ConstGenerator trait is implemented by following types.

f32

f64

num::Complex<f32>

num::Complex<f64>

bool

i32

u32

u8

i64

u64

i16

u16

Associated Types

type OutType: HasAfEnum

The type of Array object returned by generate function

Required methods

fn generate(&self, dims: Dim4) -> Array<Self::OutType>

Create an Array of dims size from scalar value self.

Parameters

dims are the dimensions of the output constant Array