Trait arrayfire::Fromf64

pub trait Fromf64 {
    fn fromf64(value: f64) -> Self;
}

Trait to convert reduction’s scalar output to appropriate output type

This is an internal trait and ideally of no use to user usecases.

Required methods

fn fromf64(value: f64) -> Self

Convert to target type from a double precision value