Trait arrayfire::ImplicitPromote

pub trait ImplicitPromote<RHS>: HasAfEnum {
    type Output: HasAfEnum;
}

This is an internal trait defined and implemented by ArrayFire create for rust’s built-in types to figure out the data type binary operation’s results.

Associated Types

type Output: HasAfEnum

This type alias points to the type of the result obtained by performing a given binary option on given type and RHS.