Crate arrayfire

Macros

join_many!
Join multiple Arrays along a given dimension

Structs

Array
A multidimensional data container

Dim4
Dim4 is used to store Array dimensions

Features
A set of Array objects (usually, used in Computer vision context)

Indexer
Struct to manage an array of resources of type af_indexer_t(ArrayFire C struct)

Seq
Sequences are used for indexing Arrays

Window
Used to render Array objects

Enums

AfError
Error codes

Aftype
Types of Array data type

BorderType
Helps determine how to pad kernels along borders

ColorMap
Dictates what color map is used for Image rendering

ColorSpace
Identify the color space of given image(Array)

Connectivity
Used by regions function to identify type of connectivity

ConvDomain
Helps determine if convolution is in Spatial or Frequency domain

ConvMode
Helps determine the size of output of convolution

InterpType
Dictates the interpolation method to be used by a function

MatProp
Helps determine the type of a Matrix

MatchType
Error metric used by matchTemplate function

NormType
Norm type

Functions

abs

accum
Reduction operation along specific dimension

acos

acosh

add

all_true
Reduction operation along specific dimension

all_true_all
Reduction operation of all values

and

any_true
Reduction operation along specific dimension

any_true_all
Reduction operation of all values

approx1
Perform signal interpolation for 1d signals

approx2
Perform signal interpolation for 2d signals

arg

asin

asinh

assign_gen
Assign an Array to another after indexing it using any combination of Array's and Sequence's

assign_seq
Assign(copy) content of an Array to another Array indexed by Sequences

atan

atan2

atanh

bilateral
Bilateral Filter.

bitand

bitor

bitxor

cbrt

ceil

cholesky
Perform Cholesky decomposition

cholesky_inplace
Perform inplace Cholesky decomposition

col
Extract col_num col from input Array

color_space
Color space conversion

cols
Get all cols from first to last in the input Array

conjg

constant
Create an Array with constant value

convolve1
Convolution

convolve2
Convolution

convolve2_sep
Separable convolution for 2d signals

convolve3
Convolution

corrcoef
Compute correlation coefficient

cos

cosh

count
Reduction operation along specific dimension

count_all
Reduction operation of all values

cov
Compute covariance of two Arrays

cplx

cplx2

det
Find the determinant of the matrix

diag_create
Create a diagonal matrix

diag_extract
Extract diagonal from a given Matrix

diff1
Reduction operation along specific dimension

diff2
Reduction operation along specific dimension

dilate

dilate3

div

dot
Calculate the dot product of vectors.

eq

erf

erfc

erode

erode3

exp

expm1

factorial

fast
Fast feature detector

fft
Fast fourier transform for 1d signals

fft2
Fast fourier transform for 2d signals

fft3
Fast fourier transform for 3d signals

fft_convolve1
Convolution using Fast-fourier transform

fft_convolve2
Convolution using Fast-fourier transform

fft_convolve3
Convolution using Fast-fourier transform

fir
Finite impulse filter

flat
Flatten the multidimensional Array to an 1D Array

flip
Flip the Array

floor

gaussian_kernel
Creates a Gaussian Kernel.

ge

get_seed
Get the seed of random number generator

get_version
Get ArrayFire Version Number

gradient
Calculate the gradients

gray2rgb
Color space conversion functions

gt

hamming_matcher
Hamming feature matcher

hist_equal
Histogram Equalization

histogram
Compute Histogram of an Array

hsv2rgb
Color space conversion functions

hypot

identity

ifft
Inverse fast fourier transform for 1d signals

ifft2
Inverse fast fourier transform for 1d signals

ifft3
Inverse fast fourier transform for 1d signals

iir
Infinite impulse response filter

imag

imax
Reduction operation along specific dimension

imax_all
Reduction operation of all values

imin
Reduction operation along specific dimension

imin_all
Reduction operation of all values

index
Indexes the input Array using seqs Sequences

index_gen
Index an Array using any combination of Array's and Sequence's

info
Print library meta-info

inverse
Compute inverse of a matrix

iota
Create a range of values

isinf

isnan

iszero

join
Join two arrays

join_many
Join multiple arrays

le

lgamma

load_image
Load Image into Array

locate
Locate the indices of non-zero elements.

log

log10

log1p

log2

lookup
Lookup(hash) an Array using another Array

lower
Create lower triangular matrix

lt

lu
Perform LU decomposition

lu_inplace
Perform inplace LU decomposition

match_template
Image matching

matmul
Matrix multiple of two Arrays

max
Reduction operation along specific dimension

max_all
Reduction operation of all values

maxfilt

maxof

mean
Compute a statistic along given dimension for an Array

mean_all
Compute statistic for all elements of Array

mean_all_weighted
Compute weighted statistic for all elements of Array

mean_shift
Meanshift Filter.

mean_weighted
Compute a weighted statistic along given dimension for an Array

medfilt

median
Compute a statistic along given dimension for an Array

median_all
Compute statistic for all elements of Array

min
Reduction operation along specific dimension

min_all
Reduction operation of all values

minfilt

minof

moddims
Change the shape of the Array

modulo

mul

neq

norm
Find the norm of a matrix

or

orb
ORB feature descriptor

pow

pow2

print
Print data in the Array

product
Reduction operation along specific dimension

product_all
Reduction operation of all values

qr
Perform QR decomposition

qr_inplace
Perform inplace QR decomposition

randn

randu

range
Create a Range of values

rank
Find rank of a matrix

real

regions
Find blobs in given image.

rem

reorder

resize
Resize an Image

rgb2gray
Color space conversion functions

rgb2hsv
Color space conversion functions

root

rotate
Rotate an Image

round

row
Extract row_num row from input Array

rows
Get all rows from first to last in the input Array

save_image
Save an Array to an image file

scale

set_col
Set col col_num in input Array to a new Array new_col

set_cols
Set cols from first to last in input Array with cols from Array new_cols

set_device
Set active device

set_intersect
Find intersection of two sets

set_row
Set row row_num in input Array to a new Array new_row

set_rows
Set rows from first to last in input Array with rows from Array new_rows

set_seed
Set seed for random number generation

set_slice
Set slice slice_num in input Array to a new Array new_slice

set_slices
Set first to last slices of input Array to a new Array new_slices

set_union
Find union of two sets

set_unique
Find unique values from a Set

shift

shiftl

shiftr

sign

sin

sinh

skew
Skew an image

slice
Get slice slice_num from input Array

slices
Get slices from first to last in input Array

sobel
Sobel Operator

solve
Solve a system of equations

solve_lu
Solve a system of equations

sort
Sort the values in input Arrays

sort_by_key
Sort the values in input Arrays

sort_index
Sort the values in input Arrays

sqrt

stdev
Compute a statistic along given dimension for an Array

stdev_all
Compute statistic for all elements of Array

sub

sum
Reduction operation along specific dimension

sum_all
Reduction operation of all values

tan

tanh

tgamma

tile

transform
Transform(Affine) an Image

translate

transpose
Transpose of a matrix.

transpose_inplace
Inplace transpose of a matrix.

trunc

upper
Create upper triangular matrix

var
Compute Variance along a specific dimension

var_all
Compute Variance of all elements

var_all_weighted
Compute weighted statistic for all elements of Array

var_weighted
Compute a weighted statistic along given dimension for an Array