Enum arrayfire::AfError

#[repr(i32)]
pub enum AfError {
    SUCCESS,
    ERR_NO_MEM,
    ERR_DRIVER,
    ERR_RUNTIME,
    ERR_INVALID_ARRAY,
    ERR_ARG,
    ERR_SIZE,
    ERR_TYPE,
    ERR_DIFF_TYPE,
    ERR_BATCH,
    ERR_DEVICE,
    ERR_NOT_SUPPORTED,
    ERR_NOT_CONFIGURED,
    ERR_NO_DBL,
    ERR_NO_GFX,
    ERR_INTERNAL,
    ERR_UNKNOWN,
}

Error codes

Variants

SUCCESS

The function returned successfully

ERR_NO_MEM

The system or device ran out of memory

ERR_DRIVER

There was an error in the device driver

ERR_RUNTIME

There was an error with the runtime environment

ERR_INVALID_ARRAY

The input array is not a valid Array object

ERR_ARG

One of the function arguments is incorrect

ERR_SIZE

The size is incorrect

ERR_TYPE

The type is not suppported by this function

ERR_DIFF_TYPE

The type of the input arrays are not compatible

ERR_BATCH

Function does not support GFOR / batch mode

ERR_DEVICE

Input does not belong to the current device

ERR_NOT_SUPPORTED

The option is not supported

ERR_NOT_CONFIGURED

This build of ArrayFire does not support this feature

ERR_NO_DBL

This device does not support double

ERR_NO_GFX

This build of ArrayFire was not built with graphics or this device does not support graphics

ERR_INTERNAL

There was an internal error either in ArrayFire or in a project upstream

ERR_UNKNOWN

Unknown Error