Enum arrayfire::NormType

pub enum NormType {
    VECTOR_1,
    VECTOR_INF,
    VECTOR_2,
    VECTOR_P,
    MATRIX_1,
    MATRIX_INF,
    MATRIX_2,
    MATRIX_L_PQ,
}

Norm type

Variants

VECTOR_1

Treats input as a vector and return sum of absolute values

VECTOR_INF

Treats input as vector and return max of absolute values

VECTOR_2

Treats input as vector and returns euclidean norm

VECTOR_P

Treats input as vector and returns the p-norm

MATRIX_1

Return the max of column sums

MATRIX_INF

Return the max of row sums

MATRIX_2

Returns the max singular value (Currently not supported)

MATRIX_L_PQ

Returns Lpq-norm