Function arrayfire::orb

pub fn orb(input: &Array, fast_thr: f32, max_feat: u32, scl_fctr: f32, levels: u32, blur_img: bool) -> (Features, Array)

ORB feature descriptor

Extract ORB descriptors from FAST features that hold higher Harris responses. FAST does not compute orientation, thus, orientation of features is calculated using the intensity centroid. As FAST is also not multi-scale enabled, a multi-scale pyramid is calculated by downsampling the input image multiple times followed by FAST feature detection on each scale.

Parameters

input - the input image Array

fast_thr - FAST threshold for which a pixel of the circle around the central pixel is considered to be brighter or darker

max_feat - maximum number of features to hold

scl_fctr - factor to downsample the input image, meaning that each level with hold prior level dimensions divided by scl_fctr

levels - number of levels to be computed for the image pyramid

blur_img - blur image with a Gaussian filter with sigma=2 before computing descriptors to increase robustness against noise if true

Return Values

This function returns a tuple of Features and Array. The features objects composed of Arrays for x and y coordinates, score, orientation and size of selected features. The Array object is a two dimensional Array of size Nx8 where N is number of selected features.