Function arrayfire::range

pub fn range<T: HasAfEnum>(dims: Dim4, seq_dim: i32) -> Array<T>

Create a Range of values

Creates an array with [0, n] values along the seq_dim which is tiled across other dimensions.

Parameters

dims is the size of Array

seq_dim is the dimension along which range values are populated, all values along other dimensions are just repeated

Return Values

Array