Function arrayfire::scale

pub fn scale(input: &Array,
             scale0: f32,
             scale1: f32,
             odim0: i64,
             odim1: i64,
             method: InterpType)
             -> Array

Scale an Image

Scale is the same functionality as resize except that the scale function uses the transform kernels. The other difference is that scale does not set boundary values to be the boundary of the input array. Instead these are set to 0.

Scale is a special case of the transform function.

Parameters

input is input image

trans0 is amount by which the first dimension is translated

trans1 is amount by which the second dimension is translated

odim0 is the first output dimension

odim1 is the second output dimension

method is the interpolation type (Nearest by default)

Return Values

Translated Image(Array).