Function arrayfire::skew

pub fn skew(input: &Array,
            skew0: f32,
            skew1: f32,
            odim0: i64,
            odim1: i64,
            method: InterpType,
            is_inverse: bool)
            -> Array

Skew an image

Skew function skews the input array along dim0 by skew0 and along dim1 by skew1. The skew areguments are in radians. Skewing the data means the data remains parallel along 1 dimensions but the other dimensions gets moved along based on the angle. If both skew0 and skew1 are specified, then the data will be skewed along both directions. Explicit output dimensions can be specified using odim0 and odim1. All new values that do not map to a location of the input array are set to 0.

Skew is a special case of the transform function.

Parameters

input is the image to be skewed

skew0 is the factor by which data is skewed along first dimension

skew1 is the factor by which data is skewed along second dimension

odim0 is the output length along first dimension

odim1 is the output length along second dimension

method indicates which interpolation method to use for rotating the image. It uses enum InterpType to identify the interpolation method.

is_inverse indicates if to apply inverse/forward transform

Return Values

Skewed Image