Function arrayfire::sobel

pub fn sobel(input: &Array, ker_size: u32) -> Result<(Array, Array), AfError>

Sobel Operator

Sobel operators perform a 2-D spatial gradient measurement on an image to emphasize the regions of high spatial frequency, namely edges. A more in depth discussion on it can be found here.

Parameters

input is the input image

ker_size is the kernel size of sobel operator

Return Values

A tuple of Arrays.

The first Array has derivatives along horizontal direction

The second Array has derivatives along vertical direction