Struct arrayfire::Array

pub struct Array {
    // some fields omitted
}

A multidimensional data container

Currently, Array objects can store only data until four dimensions

Methods

impl Array

fn new<T: HasAfEnum>(slice: &[T], dims: Dim4) -> Array

Constructs a new Array object

Examples

use arrayfire::{Array, Dim4, print};
let values: [f32; 3] = [1.0, 2.0, 3.0];
let indices = Array::new(&values, Dim4::new(&[3, 1, 1, 1]));
print(&indices);

fn new_strided<T: HasAfEnum>(slice: &[T], offset: i64, dims: Dim4, strides: Dim4) -> Array

Constructs a new Array object from strided data

The data pointed by the slice passed to this function can possibily be offseted using an additional offset parameter.

fn get_backend(&self) -> Backend

Returns the backend of the Array

Return Values

Returns an value of type Backend which indicates which backend was active when Array was created.

fn get_device_id(&self) -> i32

Returns the device identifier(integer) on which the Array was created

Return Values

Return the device id on which Array was created.

fn elements(&self) -> i64

Returns the number of elements in the Array

fn get_type(&self) -> DType

Returns the Array data type

fn dims(&self) -> Dim4

Returns the dimensions of the Array

fn strides(&self) -> Dim4

Returns the strides of the Array

fn numdims(&self) -> u32

Returns the number of dimensions of the Array

fn offset(&self) -> i64

Returns the offset to the pointer from where data begins

fn get(&self) -> i64

Returns the native FFI handle for Rust object Array

fn host<T>(&self, data: &mut [T])

Copies the data from the Array to the mutable slice data

fn eval(&self)

Evaluates any pending lazy expressions that represent the data in the Array object

fn copy(&self) -> Array

Makes an copy of the Array

Internally, this is handled by reference counting

fn is_empty(&self) -> bool

Checks if the Array is of specific format/data type.

fn is_scalar(&self) -> bool

Checks if the Array is of specific format/data type.

fn is_row(&self) -> bool

Checks if the Array is of specific format/data type.

fn is_column(&self) -> bool

Checks if the Array is of specific format/data type.

fn is_vector(&self) -> bool

Checks if the Array is of specific format/data type.

fn is_complex(&self) -> bool

Checks if the Array is of specific format/data type.

fn is_double(&self) -> bool

Checks if the Array is of specific format/data type.

fn is_single(&self) -> bool

Checks if the Array is of specific format/data type.

fn is_real(&self) -> bool

Checks if the Array is of specific format/data type.

fn is_floating(&self) -> bool

Checks if the Array is of specific format/data type.

fn is_integer(&self) -> bool

Checks if the Array is of specific format/data type.

fn is_bool(&self) -> bool

Checks if the Array is of specific format/data type.

fn is_linear(&self) -> bool

Checks if the Array is of specific format/data type.

fn is_owner(&self) -> bool

Checks if the Array is of specific format/data type.

fn cast<T: HasAfEnum>(&self) -> Array

Cast the Array data type to target_type