Struct arrayfire::Array

pub struct Array {
    // some fields omitted
}

A multidimensional data container

Currently, Array objects can store only data until four dimensions

Methods

impl Array

fn new<T>(dims: Dim4, slice: &[T], aftype: Aftype) -> Result<Array, AfError>

Constructs a new Array object

Examples

let values: &[f32] = &[1.0, 2.0, 3.0];
let indices = Array::new(Dim4::new(&[3, 1, 1, 1]), values, Aftype::F32).unwrap();

fn get_backend(&self) -> Backend

Returns the backend of the Array

Return Values

Returns an value of type Backend which indicates which backend was active when Array was created.

fn elements(&self) -> Result<i64, AfError>

Returns the number of elements in the Array

fn get_type(&self) -> Result<Aftype, AfError>

Returns the Array data type

fn dims(&self) -> Result<Dim4, AfError>

Returns the dimensions of the Array

fn numdims(&self) -> Result<u32, AfError>

Returns the number of dimensions of the Array

fn get(&self) -> i64

Returns the native FFI handle for Rust object Array

fn host<T>(&self, data: &mut [T]) -> Result<(), AfError>

Copies the data from the Array to the mutable slice data

fn eval(&self) -> Result<(), AfError>

Evaluates any pending lazy expressions that represent the data in the Array object

fn copy(&self) -> Result<Array, AfError>

Makes an copy of the Array

Internally, this is handled by reference counting

fn is_empty(&self) -> Result<bool, AfError>

Checks if the Array is of specific format/data type.

fn is_scalar(&self) -> Result<bool, AfError>

Checks if the Array is of specific format/data type.

fn is_row(&self) -> Result<bool, AfError>

Checks if the Array is of specific format/data type.

fn is_column(&self) -> Result<bool, AfError>

Checks if the Array is of specific format/data type.

fn is_vector(&self) -> Result<bool, AfError>

Checks if the Array is of specific format/data type.

fn is_complex(&self) -> Result<bool, AfError>

Checks if the Array is of specific format/data type.

fn is_double(&self) -> Result<bool, AfError>

Checks if the Array is of specific format/data type.

fn is_single(&self) -> Result<bool, AfError>

Checks if the Array is of specific format/data type.

fn is_real(&self) -> Result<bool, AfError>

Checks if the Array is of specific format/data type.

fn is_floating(&self) -> Result<bool, AfError>

Checks if the Array is of specific format/data type.

fn is_integer(&self) -> Result<bool, AfError>

Checks if the Array is of specific format/data type.

fn is_bool(&self) -> Result<bool, AfError>

Checks if the Array is of specific format/data type.

fn cast(&self, target_type: Aftype) -> Result<Array, AfError>

Cast the Array data type to target_type