Struct arrayfire::Array

pub struct Array<T: HasAfEnum> { /* fields omitted */ }

A multidimensional data container

Currently, Array objects can store only data until four dimensions

NOTE

All operators(traits) from std::ops module implemented for Array object carry out element wise operations. For example, * does multiplication of elements at corresponding locations in two different Arrays.

Methods

impl<T> Array<T> where
T: HasAfEnum,

pub fn new(slice: &[T], dims: Dim4) -> Self

Constructs a new Array object

Examples

use arrayfire::{Array, Dim4, print};
let values: [f32; 3] = [1.0, 2.0, 3.0];
let indices = Array::new(&values, Dim4::new(&[3, 1, 1, 1]));
print(&indices);

pub fn new_strided(slice: &[T], offset: i64, dims: Dim4, strides: Dim4) -> Self

Constructs a new Array object from strided data

The data pointed by the slice passed to this function can possibily be offseted using an additional offset parameter.

pub fn new_empty(dims: Dim4) -> Self

Constructs a new Array object of specified dimensions and type

Examples

use arrayfire::{Array, Dim4};
let garbageVals = Array::<f32>::new_empty(Dim4::new(&[3, 1, 1, 1]));

pub fn get_backend(&self) -> Backend

Returns the backend of the Array

Return Values

Returns an value of type Backend which indicates which backend was active when Array was created.

pub fn get_device_id(&self) -> i32

Returns the device identifier(integer) on which the Array was created

Return Values

Return the device id on which Array was created.

pub fn elements(&self) -> usize

Returns the number of elements in the Array

pub fn get_type(&self) -> DType

Returns the Array data type

pub fn dims(&self) -> Dim4

Returns the dimensions of the Array

pub fn strides(&self) -> Dim4

Returns the strides of the Array

pub fn numdims(&self) -> u32

Returns the number of dimensions of the Array

pub fn offset(&self) -> i64

Returns the offset to the pointer from where data begins

pub fn get(&self) -> i64

Returns the native FFI handle for Rust object Array

pub fn host<O: HasAfEnum>(&self, data: &mut [O])

Copies the data from the Array to the mutable slice data

pub fn eval(&self)

Evaluates any pending lazy expressions that represent the data in the Array object

pub fn copy(&self) -> Self

Makes an copy of the Array

This does a deep copy of the data into a new Array

pub fn is_empty(&self) -> bool

Check if Array is empty

pub fn is_scalar(&self) -> bool

Check if Array is scalar

pub fn is_row(&self) -> bool

Check if Array is a row

pub fn is_column(&self) -> bool

Check if Array is a column

pub fn is_vector(&self) -> bool

Check if Array is a vector

pub fn is_complex(&self) -> bool

Check if Array is of complex type

pub fn is_double(&self) -> bool

Check if Array's numerical type is of double precision

pub fn is_single(&self) -> bool

Check if Array's numerical type is of single precision

pub fn is_real(&self) -> bool

Check if Array is of real type

pub fn is_floating(&self) -> bool

Check if Array is of single precision

pub fn is_integer(&self) -> bool

Check if Array is of integral type

pub fn is_bool(&self) -> bool

Check if Array is of boolean type

pub fn is_linear(&self) -> bool

Check if Array's memory layout is continuous and one dimensional

pub fn is_owner(&self) -> bool

Check if Array's memory is owned by it and not a view of another Array

pub fn cast<O: HasAfEnum>(&self) -> Array<O>

Cast the Array data type to target_type

pub fn is_sparse(&self) -> bool

Find if the current array is sparse

pub fn lock(&self)

Lock the device buffer in the memory manager

Locked buffers are not freed by memory manager until unlock is called.

pub fn unlock(&self)

Unlock the device buffer in the memory manager

This function will give back the control over the device pointer to the memory manager.

pub fn device_ptr(&self) -> u64

Get the device pointer and lock the buffer in memory manager

The device pointer is not freed by memory manager until unlock is called.

pub fn get_allocated_bytes(&self) -> usize

Get the size of physical allocated bytes.

This function will return the size of the parent/owner if the current Array object is an indexed Array.