Struct arrayfire::Indexer

pub struct Indexer { /* fields omitted */ }

Struct to manage an array of resources of type af_indexer_t(ArrayFire C struct)

Methods

impl Indexer

fn new() -> Indexer

fn set_index<T: Indexable>(&mut self, idx: &T, dim: u32, is_batch: Option<bool>)

Set either Array or Seq to index an Array along idx dimension

fn get(&self) -> i64

Get native(ArrayFire) resource handle

fn len(&self) -> u32

Get number of indexers

This can be a maximum of four since currently ArrayFire supports maximum of four dimensions