Struct arrayfire::Window

pub struct Window {
    // some fields omitted
}

Used to render Array objects

The renderings can be either plots, histograms or simply just image displays. A single window can also display multiple of the above renderings at the same time, which is known as multiview mode. An example of that is given below.

Examples

let wnd = Window::new(1280, 720, String::from("Image Histogram")).unwrap();
let img = match load_image("Path to image", true/*If color image, 'false' otherwise*/) {
    Ok(img) => img,
    Err(err) => panic!("Image loading failed with error code {}", err),
};
let hst = histogram(img, 256, 0, 255).unwrap();

loop {
    wnd.grid(2, 1);

    wnd.set_view(0, 0);
    wnd.draw_image(img, Some("Input Image"));

    wnd.set_view(1, 0);
    wnd.draw_histogram(hst, 0.0, 255.0, Some("Input Image Histogram"));

    wnd.show();

    if wnd.is_closed().unwrap() == true { break; }
}

Methods

impl Window

fn new(width: i32, height: i32, title: String) -> Result<Window, AfError>

Creates new Window object

fn set_position(&self, x: u32, y: u32) -> Result<(), AfError>

Set window starting position on the screen

fn set_title(&self, title: String) -> Result<(), AfError>

Set window title

fn set_colormap(&mut self, cmap: ColorMap)

Set color map to be used for rendering image, it can take one of the values of enum ColorMap

fn is_closed(&self) -> Result<bool, AfError>

Returns true if the window close is triggered by the user

fn grid(&self, rows: i32, cols: i32) -> Result<(), AfError>

Used to setup display layout in multiview mode

fn show(&mut self) -> Result<(), AfError>

Used in multiview mode to swap back buffer with front buffer to show the recently rendered frame

fn set_view(&mut self, r: i32, c: i32)

Used in multiview mode to set the current sub-region to which the subsequence draw call renders to

fn draw_image(&self, input: &Array, title: Option<String>)

Render given Array as an image

fn draw_plot(&self, x: &Array, y: &Array, title: Option<String>)

Render given two Array's x and y as a 2d line plot

fn draw_hist(&self, hst: &Array, minval: f64, maxval: f64, title: Option<String>)

Render given Array as a histogram