Trait arrayfire::ConstGenerator

pub trait ConstGenerator {
    fn generate(&self, dims: Dim4) -> Array;
}

Type Trait to generate a constant Array of given size

Internally, ConstGenerator trait is implemented by following types.

f32

f64

num::Complex<f32>

num::Complex<f64>

bool

i32

u32

u8

i64

u64

i16

u16

Required Methods

fn generate(&self, dims: Dim4) -> Array

Create an Array of dims size from scalar value self.

Parameters

dims are the dimensions of the output constant Array