Trait arrayfire::FloatingPoint

pub trait FloatingPoint: HasAfEnum {
    fn is_real() -> bool { ... }

    fn is_complex() -> bool { ... }
}

Trait qualifier to accept either real or complex typed data

Provided methods

fn is_real() -> bool

Use to check if trait implementor is real number

fn is_complex() -> bool

Use to check if trait implementor is complex number

Loading content...