Trait arrayfire::Indexable

pub trait Indexable {
    fn set(&self, idxr: &mut Indexer, dim: u32, is_batch: Option<bool>);
}

Trait bound indicating indexability

Any object to be able to be passed on to Indexer::set_index() method should implement this trait with appropriate implementation of set method.

Required Methods

fn set(&self, idxr: &mut Indexer, dim: u32, is_batch: Option<bool>)

Set indexing object for a given dimension

Parameters

idxr is mutable reference to Indexer object which will be modified to set self indexable along dim dimension.

dim is the dimension along which self indexable will be used for indexing.

is_batch is only used if self is Seq to indicate if indexing along dim is a batched operation.